GPLIM-834, GPLIM-805: Handle case where user clicks on delete/abandon OK button more than once before the dialog is closed.
Since sample position changes every time the sample list changes, it is not a good way to track samples, so also changed the code to use the sample's database ID instead. Also removed the sample position from the message in the JIRA ticket for abandon operations.

This fixes the On Risk dialog OK button too.

Based on discussions with hrafal, also removed equals() and hashCode() for ProductOrderSample, so this class will now use the default hibernate provided methods.