GPLIM-833: Remove use of @Named in Mercury.
The @Named identifier is no longer necessary since we are no longer using JSF or instantiating beans by name in our code.

Updated code to use function calls in a few places when generating URLs for BSP or JIRA; added a helper method in the action bean base class to create JIRA URLs.