GPLIM-1: Update PDO sample count code as per suggestions in code review GPI-139.
The original change backed out a change where sample count was no longer being displayed correctly for draft PDOs. This regression occurred as part of the % Complete display work, because that code was using a query that excluded Draft PDOs, and didn't have an easy way to access Draft PDOs in the % Complete results.

The original code worked, but it duplicated the work to count the number of all samples in a PDO. It was also computing this total every time a PDO was fetched, even if the value wasn't used.

The new code adds support for fetching draft results and storing the results. It also better encapsulates the way Draft PDOs are annotated.