GPLIM-1876: Fix issue where the total sample count for a draft PDO would show up as zero. This count is shown in the PDO list page, the RP list of PDOs in the RP detail page, and is used to draw the progress bar on the PDO details page.
This appears to have been broken when the Search capability was added to the PDO list page. At this time the sample progress data was limited to the list of PDOs shown; prior to this the data for all PDOs including ones not shown was fetched.

This is fixed by using a list of PDO database IDs instead of JIRA keys to search for the progress data.