GPLIM-2038: Fix issue where workflow visualizations wouldn't be shown if they weren't already cached.
The code was checking to see if the visualization image was already generated before generating the image URL. However this would avoid calling into the web service that would generate the image in the first place. So the code would only work in the case where the images were already generated, most likely via calling the web service directly while testing.