GPLIM-1: Fix regression due to change in layout.jsp
layout.jsp was changed so that it no longer instantiates its own action bean, which caused a bad redirect (a redirect without context) to fail. Fixing the redirect to use the action bean class fixed the problem. All other uses of RedirectResolution also use the action bean class, not the JSP, so this should be the only case of this problem.