Revert "GPLIM-2885 GPLIM-2887 Row numbers for the other locations where duplicates may be found, still need to put row numbers on the current manifest record being analyzed."
Temporary revert, these changes cause multiple dbfree tests to fail.  The new logic is not broken, but the tests are fragile and are looking for specific exception texts
which are no longer valid.  The tests and possibly production code should be modified to confirm that the expected conditions are being detected.

This reverts commit e78d3d6a57daf3c15e022c858c1230331190b9a4.