PORTAL-221: Fix issue with test where it was hardcoding the display text for the BSP Client MaterialInfo enum.
The display text for an instance of MaterialInfo was changed to fix a spelling error, which caused a test to fail since it used a hardcoded string that also had the spelling error. The test now uses the enum itself to get the display test, which will be compatible with future non-API changes to the enum. Mercury itself does not look up enums using hardcoded display text, so this test case didn't represent and actual use case that could be affected by this change to the BSP Client library.